Update ValueInc_project1.py
Value Inc is a retail store that sells household items all over the world by bulk.
The Sales Manager has no sales reporting but he has a brief idea of current sales.
He also has no idea of the monthly cost, profit and top selling products. He wants a
dashboard on this and says the data is currently stored in an excel sheet.