BlogMe.py
BlogMe, a famous blogging business has a dataset of news articles that they need
further analysis on.
Firstly, they’d like keywords to be extracted from headlines of the article. Secondly,
they would need to determine the sentiment of the news articles. The data is in an
excel sheet and they would like to see a dashboard outlying sentiment, top articles etc.